requirements.txt
asgiref==3.3.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
Pillow==8.1.0
psycopg2==2.8.6
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
